update sourmash to r0.17.0

+++ src/fastmultigather.rs
@@ -29,7 +29,7 @@ use crate::utils::{
 pub fn fastmultigather(
     query_filepath: String,
     against_filepath: String,
-    threshold_bp: usize,
+    threshold_bp: u32,
     scaled: Option<usize>,
     selection: Selection,
     allow_failed_sigpaths: bool,
@@ -47,9 +47,9 @@ pub fn fastmultigather(
     )?;
 
     let scaled = match scaled {
-        Some(s) => s,
+        Some(s) => s as u32,
         None => {
-            let scaled = *query_collection.max_scaled().expect("no records!?") as usize;
+            let scaled = *query_collection.max_scaled().expect("no records!?");
             eprintln!(
                 "Setting scaled={} based on max scaled in query collection",
                 scaled
@@ -164,7 +164,7 @@ pub fn fastmultigather(
                         query_name,
                         query_filename,
                         query_mh,
-                        scaled as u64,
+                        scaled,
                         matchlist,
                         threshold_hashes,
                         Some(gather_output),
